refactor: Rename remote-editor to editor-loader
The module focuses on loading editor assets.
