+++ src/utils/editor-loader.js
